Move bench tests to record_bench_test.go
